Created Directory and moved my first project
