--- main.py
@@ -1,23 +1,0 @@
-# from turtle import Screen
-from snake import Snake
-from food import Food
-
-# snack = Snake()
-
-snack = Snake(1000, 500)
-
-# pos = [0, -20, -40]
-# body = []
-# for i in range(0, 3):
-#     pixel = Turtle("square")
-#     pixel.color("white")
-#     pixel.penup()
-#     pixel.goto(pos[i], 0)
-#     body.append(pixel)
-#
-# game_is_on = True
-# while game_is_on:
-#     for i in body:
-#     # while i.xcor() < 250:
-#         i.forward(10)
-
